+++ Source/StatsGenerator/Presets/DefaultStatsPreset.h
@@ -51,11 +51,11 @@ class DefaultStatsPreset : public AbstractMicrostructurePreset
     virtual ~DefaultStatsPreset();
 
 
-    void generateOmega3Data(StatsGenPlotWidget* plot, QVector<double> binNumbers);
-    void generateBOverAPlotData(StatsGenPlotWidget* plot, QVector<double> binNumbers);
-    void generateCOverAPlotData(StatsGenPlotWidget* plot, QVector<double> binNumbers);
-    void generateCOverBPlotData(StatsGenPlotWidget* plot, QVector<double> binNumbers);
-    void generateNeighborPlotData(StatsGenPlotWidget* plot, QVector<double> binNumbers);
+    void generateOmega3Data(StatsGenPlotWidget* plot, QVector<float> binNumbers);
+    void generateBOverAPlotData(StatsGenPlotWidget* plot, QVector<float> binNumbers);
+    void generateCOverAPlotData(StatsGenPlotWidget* plot, QVector<float> binNumbers);
+    void generateCOverBPlotData(StatsGenPlotWidget* plot, QVector<float> binNumbers);
+    void generateNeighborPlotData(StatsGenPlotWidget* plot, QVector<float> binNumbers);
 
   protected:
     DefaultStatsPreset();
